refactor: move TablePlugin to components and fix imports
- Move TablePlugin.tsx from src/plugins/ to src/components/ to match
  CodeBlockPlugin.tsx convention
- Import INSERT_TABLE_COMMAND directly from @lexical/table in
  SlashCommandMenu.tsx instead of re-exporting through TablePlugin
- Update Editor.tsx import path accordingly
